requirements.txt
pywebview>=4.0.0
pyinstaller>=5.0.0
Pillow>=10.0.0
